feat: adicionando regra para deletar autor

+++ src/main/java/com/projeto/libraryapi/service/AutorService.java
@@ -1,7 +1,9 @@
 package com.projeto.libraryapi.service;
 
+import com.projeto.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.projeto.libraryapi.model.Autor;
 import com.projeto.libraryapi.repository.AutorRepository;
+import com.projeto.libraryapi.repository.LivroRepository;
 import com.projeto.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -14,11 +16,16 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository autorRepository;
+    private final LivroRepository livroRepository;
     private final AutorValidator autorValidator;
 
-    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator) {
+    public AutorService(
+            AutorRepository autorRepository,
+            AutorValidator autorValidator,
+            LivroRepository livroRepository) {
         this.autorRepository = autorRepository;
         this.autorValidator = autorValidator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
@@ -28,7 +35,8 @@ public class AutorService {
 
     public void atualizar(Autor autor) {
         if(Objects.isNull(autor.getId())) {
-            throw new IllegalArgumentException("Para atualizar, é necessário que o autor já esteja salvo na base.");
+            throw new IllegalArgumentException(
+                    "Para atualizar, é necessário que o autor já esteja salvo na base.");
         }
         autorValidator.validar(autor);
         autorRepository.save(autor);
@@ -39,7 +47,15 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um Autor que possui livros cadastrados!");
+        }
         autorRepository.delete(autor);
+    }
+
+    private boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 
     public List<Autor> pesquisar(String nome, String nacionalidade) {
